Day 3.1 wip. Improve serial protocol.
It was getting stuck on incomplete packets.
I'm sure i'm inventing some stone age version of protocols in a clueless way here.
Pretty fun.

Need to fix the receiving on serial too, it waits for EOL chars for log messages,
but panics do not frame printlns with them, so they should be removed.